Decouple cy and pf cypress tests
