+++ frontend/cypress/integration/featureFiles/graph_toolbar-cy.feature
